First create package utils in repairnator-core and move the class Utils into it
