+++ repairnator/repairnator-checkbranches/src/main/java/fr/inria/spirals/repairnator/checkbranches/RunnablePipelineContainer.java
@@ -6,7 +6,7 @@ import com.spotify.docker.client.messages.ContainerConfig;
 import com.spotify.docker.client.messages.ContainerCreation;
 import com.spotify.docker.client.messages.ContainerExit;
 import com.spotify.docker.client.messages.HostConfig;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.config.RepairnatorConfig;
 
 import fr.inria.spirals.repairnator.states.LauncherMode;
