+++ repairnator/repairnator-core/src/main/java/fr/inria/spirals/repairnator/notifier/EndProcessNotifier.java
@@ -1,6 +1,6 @@
 package fr.inria.spirals.repairnator.notifier;
 
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.notifier.engines.NotifierEngine;
 
 import java.util.Date;
