+++ repairnator/repairnator-core/src/test/java/fr/inria/spirals/repairnator/utils/TestUtils.java
@@ -1,4 +1,4 @@
-package fr.inria.spirals.repairnator;
+package fr.inria.spirals.repairnator.utils;
 
 import org.junit.Test;
