+++ repairnator/repairnator-dockerpool/src/main/java/fr/inria/spirals/repairnator/dockerpool/serializer/TreatedBuildTracking.java
@@ -3,7 +3,7 @@ package fr.inria.spirals.repairnator.dockerpool.serializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import fr.inria.spirals.repairnator.serializer.ProcessSerializer;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.serializer.SerializerType;
 import fr.inria.spirals.repairnator.serializer.engines.SerializedData;
 import fr.inria.spirals.repairnator.serializer.engines.SerializerEngine;
