+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/notifier/BugAndFixerBuildsNotifier.java
@@ -2,7 +2,7 @@ package fr.inria.spirals.repairnator.notifier;
 
 import fr.inria.jtravis.entities.Build;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector4Bears;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.notifier.engines.NotifierEngine;
 import fr.inria.spirals.repairnator.process.inspectors.JobStatus;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
