+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/step/CloneRepository.java
@@ -1,7 +1,7 @@
 package fr.inria.spirals.repairnator.process.step;
 
 import fr.inria.jtravis.entities.Build;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.states.PipelineState;
 import org.eclipse.jgit.api.Git;
