+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/step/paths/ComputeModules.java
@@ -1,6 +1,6 @@
 package fr.inria.spirals.repairnator.process.step.paths;
 
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.process.step.StepStatus;
 import fr.inria.spirals.repairnator.process.maven.MavenHelper;
