+++ repairnator/repairnator-pipeline/src/test/java/fr/inria/spirals/repairnator/process/step/push/TestCommitPatch.java
@@ -3,7 +3,7 @@ package fr.inria.spirals.repairnator.process.step.push;
 import ch.qos.logback.classic.Level;
 import fr.inria.jtravis.entities.Build;
 import fr.inria.spirals.repairnator.BuildToBeInspected;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.config.RepairnatorConfig;
 import fr.inria.spirals.repairnator.process.files.FileHelper;
 import fr.inria.spirals.repairnator.process.inspectors.JobStatus;
