+++ repairnator/repairnator-realtime/src/main/java/fr/inria/spirals/repairnator/realtime/serializer/BlacklistedSerializer.java
@@ -2,7 +2,7 @@ package fr.inria.spirals.repairnator.realtime.serializer;
 
 import com.google.gson.JsonObject;
 import fr.inria.jtravis.entities.Repository;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.realtime.RTScanner;
 import fr.inria.spirals.repairnator.serializer.Serializer;
 import fr.inria.spirals.repairnator.serializer.SerializerType;
