+++ repairnator/repairnator-scanner/src/main/java/fr/inria/spirals/repairnator/serializer/ScannerSerializer.java
@@ -2,7 +2,7 @@ package fr.inria.spirals.repairnator.serializer;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import fr.inria.spirals.repairnator.Utils;
+import fr.inria.spirals.repairnator.utils.Utils;
 import fr.inria.spirals.repairnator.scanner.ProjectScanner;
 import fr.inria.spirals.repairnator.serializer.engines.SerializedData;
 import fr.inria.spirals.repairnator.serializer.engines.SerializerEngine;
